requirements/base.txt
dj-database-url==0.3.0
Django==1.9.1
django-model-utils==2.4
django-polymorphic==0.8.1
djangorestframework==3.3.2
djangorestframework-expander==0.1.2
djangorestframework-timed-auth-token==1.2.0
Pillow==3.1.0
psycopg2==2.6.1
